refactor(nat-lab): move Runtime/Events out of telio.py into runtime.py (LLT-5791)
Runtime (event/state bookkeeping) and Events (event-wait wrappers) are not the
libtelio client; they just back it. Move Runtime, Events and WontHappenError to a
new tests/runtime.py. telio.py imports them back; the only external importer
(telio_test.py) imports from tests.runtime now. Tests reach them via
client.get_runtime()/get_events() as before — no Client API change.

telio.py drops from 1161 to 785 lines.
